+++ flutter_async_app/android/app/src/main/kotlin/com/example/flutter_async_app/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.flutter_async_app
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}