+++ lib/flutter_async.dart
@@ -1,9 +1,14 @@
 library flutter_async;
 
-export 'src/widgets/async_button/async_button_base.dart';
-export 'src/async_extension.dart';
-export 'src/widgets/async_button/async_button.dart';
-export 'src/widgets/async_builder/async_builder.dart';
 export 'src/async_controller.dart';
-export 'src/widgets/async/inherited_async.dart';
+export 'src/async_extension.dart';
+export 'src/listenables/async_listenable.dart';
+export 'src/listenables/async_notifier.dart';
 export 'src/widgets/async/async_config.dart';
+export 'src/widgets/async/inherited_async.dart';
+export 'src/widgets/async_builder/async_builder.dart';
+export 'src/widgets/async_button/async_button.dart';
+export 'src/widgets/async_button/async_button_base.dart';
+
+typedef DataChanged<T> = void Function(T data);
+typedef ErrorCallback = void Function(Object error, StackTrace stackTrace);